Refactored MapTileError in separate module
Before was placed in "layer" module, but was used beyond the scope of
just layers.
